- Update permission + Logic P2
